refactor/ make package as fetch
